backend/eschool/requirements.txt
# Django and Django REST Framework
Django==5.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter (for production)
# sqlite3 is included with Python (for development)

# CORS handling
django-cors-headers==4.3.1

# Filtering and search
django-filter==23.5

# Environment variables
python-decouple==3.8

# Date and time utilities
python-dateutil==2.8.2

# Image processing (if needed for student photos)
Pillow==10.1.0

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# API documentation
drf-spectacular==0.26.5

# Production server
gunicorn==21.2.0

# Security
django-ratelimit==4.1.0

# Monitoring and logging
sentry-sdk==1.38.0

# Data validation
marshmallow==3.20.1

# Utilities
requests==2.31.0
celery==5.3.4  # For background tasks
redis==5.0.1   # For Celery broker



